+++ Emuera/Runtime/Script/Statements/FunctionArgType.cs
@@ -92,4 +92,5 @@ enum FunctionArgType
 	#endregion
 	SP_SETBGIMAGE,
 	SP_SETIMAGELAYER,
+	SP_SETIMAGELAYERL,
 }