+++ Emuera/Runtime/Utils/EvilMask/ConsoleDivPart.cs
@@ -16,10 +16,11 @@ class ConsoleDivPart : AConsoleDisplayNode
 {
 	public ConsoleDivPart(MixedNum xPos, MixedNum yPos, MixedNum width, MixedNum height, int depth, int color, StyledBoxModel box, bool isRelative, DisplayMode displayMode, ConsoleDisplayLine[] childs)
 	{
-		backgroundColor = color >= 0 ? Color.FromArgb((int)(color | 0xff000000)) : Color.Transparent;
+		backgroundColor = color != -1 ? Color.FromArgb(color) : Color.Transparent;
 		StringBuilder sb = new();
 		width.num = Math.Abs(width.num);
-		height.num = Math.Abs(height.num);
+		if (height != null)
+			height.num = Math.Abs(height.num);
 		sb.Append("<div");
 		AddTagMixedNumArg(sb, "xpos", xPos);
 		AddTagMixedNumArg(sb, "ypos", yPos);
@@ -40,7 +41,7 @@ class ConsoleDivPart : AConsoleDisplayNode
 			{
 				borderColors = new Color[4];
 				for (int i = 0; i < 4; i++)
-					borderColors[i] = box.color[i] >= 0 ? Color.FromArgb((int)(box.color[i] | 0xff000000)) : Color.Transparent;
+					borderColors[i] = box.color[i] != -1 ? Color.FromArgb(box.color[i]) : Color.Transparent;
 				AddColorParam4(sb, "bcolor", borderColors);
 			}
 		}
@@ -62,11 +63,23 @@ class ConsoleDivPart : AConsoleDisplayNode
 		#endregion
 
 		this.width = MixedNum.ToPixel(width, 0);
-		Height = MixedNum.ToPixel(height, 0);
 		children = childs;
 		Depth = depth;
 		IsRelative = isRelative;
 		Display = displayMode;
+
+		// height auto: calculate from content lines + padding/border
+		if (height != null)
+		{
+			Height = MixedNum.ToPixel(height, 0);
+		}
+		else
+		{
+			int padTop = 0, padBottom = 0;
+			if (padding != null) { padTop = padding[Direction.Top]; padBottom = padding[Direction.Bottom]; }
+			if (border != null) { padTop += border[Direction.Top]; padBottom += border[Direction.Bottom]; }
+			Height = childs.Length * Config.Config.LineHeight + padTop + padBottom;
+		}
 		
 		ShiftChildrenX(PointX + xOffset + divXOffset);
 	}